+++ clang-tools-extra/test/clang-tidy/checkers/modernize/use-equals-delete-macros.cpp
@@ -1,5 +1,5 @@
 // RUN: %check_clang_tidy %s modernize-use-equals-delete %t -- \
-// RUN:   -config="{CheckOptions: [{key: modernize-use-equals-delete.IgnoreMacros, value: false}]}"
+// RUN:   -config="{CheckOptions: {modernize-use-equals-delete.IgnoreMacros: false}}"
 
 #define MACRO(type) void operator=(type const &)
 class C {